go/2024/day08_2/day08_test.go
package main

import (
	_ "embed"
	"testing"
)

//go:embed puzzle1_test.txt
var inputTest string

func TestPart1(t *testing.T) {
	result := part1(inputTest)
	expected := 14
	if result != expected {
		t.Errorf("Result is incorrect, got: %d, want: %d.", result, expected)
	}
}

func TestPart2(t *testing.T) {
	result := part2(inputTest)
	expected := 34
	if result != expected {
		t.Errorf("Result is incorrect, got: %d, want: %d.", result, expected)
	}
}

func TestPart1Input(t *testing.T) {
	result := part1(puzzle1)
	expected := 371
	if result != expected {
		t.Errorf("Result is incorrect, got: %d, want: %d.", result, expected)
	}
}

func TestPart2Input(t *testing.T) {
	result := part2(puzzle1)
	expected := 1229
	if result != expected {
		t.Errorf("Result is incorrect, got: %d, want: %d.", result, expected)
	}
}

func BenchmarkPart1(b *testing.B) {
	for i := 0; i < b.N; i++ {
		part1(puzzle1)
	}
}
func BenchmarkPart2(b *testing.B) {
	for i := 0; i < b.N; i++ {
		part2(puzzle1)
	}
}
